time: prevent negative times in NewTime
When we read HHDD (or YYMMDD) values in ACH of 0000 (or 000000) these
get converted into 0000-00-00 at 00:00 in UTC. After converting to
America/New_York (back a few hours) this creates a "negative time",
which is completely invalid. Instead let's keep a zero time but force
it into America/New_York.