use separate DB connection for migrator to release opened connection
We pass existing DB instance to migrator (WithInstance method). Migrator
opens connection to the DB instance, runs migrations and ... never
closes it. As closing connection (calling method Close) will will close
DB instance as well. It leads to the situation when we have always 1
opened connection.

In order to solve this we create two separate DB instances. One for our
service and one for the migrator.